# aadhar_swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "AadharService",
    "version": "1.0"
  },
  "paths": {
    "/api/v1/verify/{aadharNumber}": {
      "get": {
        "tags": [
          "Aadhar"
        ],
        "parameters": [
          {
            "name": "aadharNumber",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AadharVerificationResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AadharVerificationResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AadharVerificationResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/verify": {
      "post": {
        "tags": [
          "Aadhar"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AadharVerificationRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AadharVerificationRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AadharVerificationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AadharVerificationResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AadharVerificationResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AadharVerificationResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/valid-numbers": {
      "get": {
        "tags": [
          "Aadhar"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/": {
      "get": {
        "tags": [
          "AadharService"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StringString<>f__AnonymousType4"
                }
              }
            }
          }
        }
      }
    },
    "/health": {
      "get": {
        "tags": [
          "Health"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AadharVerificationRequest": {
        "required": [
          "aadhar_number"
        ],
        "type": "object",
        "properties": {
          "aadhar_number": {
            "maxLength": 12,
            "minLength": 12,
            "pattern": "^[0-9]+$",
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "AadharVerificationResponse": {
        "type": "object",
        "properties": {
          "aadhar_number": {
            "type": "string",
            "nullable": true
          },
          "is_valid": {
            "type": "boolean"
          },
          "status": {
            "type": "string",
            "nullable": true
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "mobile_no": {
            "type": "string",
            "nullable": true
          },
          "address": {
            "type": "string",
            "nullable": true
          },
          "gender": {
            "type": "string",
            "nullable": true
          },
          "date_of_birth": {
            "type": "string",
            "nullable": true
          },
          "photo_url": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "StringString<>f__AnonymousType4": {
        "type": "object",
        "properties": {
          "service": {
            "type": "string",
            "nullable": true
          },
          "version": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    }
  }
}